requirements.txt
fastapi
uvicorn
python-docx
google-generativeai
pydantic
python-multipart